@@ -1,76 +1,81 @@
 ﻿using System.Numerics;
-using System.Runtime.InteropServices;
 using CelSerEngine.Core.Extensions;
 using CelSerEngine.Core.Models;
 
-namespace CelSerEngine.Core.Comparators
+namespace CelSerEngine.Core.Comparators;
+
+public class VectorComparer<T> : IScanComparer where T : struct, INumber<T>
 {
-    public class VectorComparer<T> : IScanComparer where T : struct
+    private readonly ScanConstraint _scanConstraint;
+    private readonly Vector<T> _userInputAsVector;
+    private readonly int _sizeOfT;
+
+    public VectorComparer(ScanConstraint scanConstraint)
     {
-        private readonly ScanConstraint _scanConstraint;
-        private readonly Vector<T> _userInputAsVector;
-        private readonly int _sizeOfT;
+        _scanConstraint = scanConstraint;
+        _userInputAsVector = new Vector<T>(scanConstraint.UserInput.ParseNumber<T>());
+        _sizeOfT = scanConstraint.ScanDataType.GetPrimitiveSize();
+    }
 
-
-        public VectorComparer(ScanConstraint scanConstraint)
+    public Vector<byte> CompareTo(ReadOnlySpan<byte> bytes)
+    {
+        return _scanConstraint.ScanCompareType switch
         {
-            _scanConstraint = scanConstraint;
-            _userInputAsVector = new Vector<T>((T)scanConstraint.UserInput);
-            _sizeOfT = Marshal.SizeOf(default(T));
-        }
+            ScanCompareType.ExactValue => Vector.AsVectorByte(Vector.Equals(new Vector<T>(bytes), _userInputAsVector)),
+            ScanCompareType.SmallerThan => Vector.AsVectorByte(Vector.LessThan(new Vector<T>(bytes), _userInputAsVector)),
+            ScanCompareType.BiggerThan => Vector.AsVectorByte(Vector.GreaterThan(new Vector<T>(bytes), _userInputAsVector)),
+            _ => throw new NotImplementedException("Not implemented")
+        };
+    }
 
-        public int GetVectorSize()
-        {
-            return Vector<T>.Count;
-        }
+    public IList<IMemorySegment> GetMatchingMemorySegments(IList<VirtualMemoryRegion> virtualMemoryRegions, IProgress<float>? progressBarUpdater = null)
+    {
+        var matchingProcessMemories = new List<IMemorySegment>();
 
-        public Vector<byte> CompareTo(ReadOnlySpan<byte> bytes)
+        for (var regionIndex = 0; regionIndex < virtualMemoryRegions.Count; regionIndex++)
         {
-            return _scanConstraint.ScanCompareType switch
+            var virtualMemoryRegion = virtualMemoryRegions[regionIndex];
+            var foundMatchingSegments = FindMatchingMemorySegmentsInRegion(virtualMemoryRegion);
+            matchingProcessMemories.AddRange(foundMatchingSegments);
+
+            if (progressBarUpdater != null)
             {
-                ScanCompareType.ExactValue => Vector.AsVectorByte(Vector.Equals(new Vector<T>(bytes), _userInputAsVector)),
-                ScanCompareType.SmallerThan => Vector.AsVectorByte(Vector.LessThan(new Vector<T>(bytes), _userInputAsVector)),
-                ScanCompareType.BiggerThan => Vector.AsVectorByte(Vector.GreaterThan(new Vector<T>(bytes), _userInputAsVector)),
-                _ => throw new NotImplementedException("Not implemented")
-            };
-        }
-
-        public IEnumerable<ProcessMemory> GetMatchingValueAddresses(IList<VirtualMemoryPage> virtualMemoryPages, IProgress<float> progressBarUpdater)
-        {
-            for (var pageIndex = 0; pageIndex < virtualMemoryPages.Count; pageIndex++)
-            {
-                var virtualMemoryPage = virtualMemoryPages[pageIndex];
-                var remaining = (int)virtualMemoryPage.RegionSize % GetVectorSize();
-
-                for (var i = 0; i < (int)virtualMemoryPage.RegionSize - remaining; i += Vector<byte>.Count)
-                {
-                    var splitBuffer = virtualMemoryPage.Bytes.AsSpan().Slice(i, Vector<byte>.Count);
-                    var compareResult = CompareTo(splitBuffer);
-
-                    if (!compareResult.Equals(Vector<byte>.Zero))
-                    {
-                        var desti = new byte[Vector<byte>.Count];
-                        Vector.AsVectorByte(compareResult).CopyTo(desti);
-                        for (var j = 0; j < Vector<byte>.Count; j += _sizeOfT)
-                        {
-                            if (compareResult[j] != 0)
-                            {
-                                var newIntPtr = (IntPtr)virtualMemoryPage.BaseAddress + i + j;
-                                var myArry = virtualMemoryPage.Bytes.AsSpan().Slice(j + i, _sizeOfT).ToArray();
-
-                                yield return 
-                                    new ProcessMemory(
-                                        virtualMemoryPage.BaseAddress, i + j,
-                                        myArry.ByteArrayToObject(_scanConstraint.ScanDataType),
-                                        _scanConstraint.ScanDataType);
-                            }
-                        }
-                    }
-                }
-
-                var progress = (float)pageIndex * 100 / virtualMemoryPages.Count;
-                progressBarUpdater?.Report(progress);
+                var progress = (float)regionIndex * 100 / virtualMemoryRegions.Count;
+                progressBarUpdater.Report(progress);
             }
+        }
+
+        return matchingProcessMemories;
+    }
+
+    private IEnumerable<IMemorySegment> FindMatchingMemorySegmentsInRegion(VirtualMemoryRegion virtualMemoryRegion)
+    {
+        var remaining = (int)virtualMemoryRegion.RegionSize % Vector<byte>.Count;
+        var regionBytes = virtualMemoryRegion.Bytes;
+
+        for (var i = 0; i < (int)virtualMemoryRegion.RegionSize - remaining; i += Vector<byte>.Count)
+        {
+            var splitBuffer = regionBytes.AsSpan().Slice(i, Vector<byte>.Count);
+            var compareResult = CompareTo(splitBuffer);
+
+            foreach (var vectorIndex in FindMatchingVectorIndexes(compareResult))
+            {
+                var offset = i + vectorIndex;
+                var memoryValue = regionBytes.AsSpan().Slice(offset, _sizeOfT).ConvertToString(_scanConstraint.ScanDataType);
+                yield return new MemorySegment(virtualMemoryRegion.BaseAddress, offset, memoryValue, _scanConstraint.ScanDataType);
+            }
+        }
+    }
+
+    private IEnumerable<int> FindMatchingVectorIndexes(Vector<byte> vectorResult)
+    {
+        if (vectorResult.Equals(Vector<byte>.Zero))
+            yield break;
+
+        for (var j = 0; j < Vector<byte>.Count; j += _sizeOfT)
+        {
+            if (vectorResult[j] != 0)
+                yield return j;
         }
     }
 }
