@@ -1,17 +1,14 @@
 ﻿using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Input;
 
-namespace CelSerEngine.Views
+namespace CelSerEngine.Wpf.Views;
+
+/// <summary>
+/// Interaction logic for MainWindow.xaml
+/// </summary>
+public partial class MainWindow : Window
 {
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
+    public MainWindow()
     {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
     }
 }
