@@ -1,12 +1,13 @@
 ﻿using CelSerEngine.Core.Extensions;
 using CelSerEngine.Core.Models;
+using System.Numerics;
 
 namespace CelSerEngine.Core.Comparators;
 
 public class ValueComparer : IScanComparer
 {
     private readonly ScanConstraint _scanConstraint;
-    private dynamic _userInput;
+    private string _userInput;
     private readonly int _sizeOfT;
 
     public ValueComparer(ScanConstraint scanConstraint)
@@ -16,15 +17,32 @@ public class ValueComparer : IScanComparer
         _sizeOfT = scanConstraint.ScanDataType.GetPrimitiveSize();
     }
 
-    public static bool CompareDataByScanConstraintType(dynamic lhs, dynamic rhs, ScanCompareType scanConstraintType)
+    public static bool MeetsTheScanConstraint(string lhs, string rhs, ScanConstraint scanConstraint)
     {
-        if (!((Type)lhs.GetType()).IsValueType)
-            throw new ArgumentException("lhs must be a ValueType (struct)");
+        return scanConstraint.ScanDataType switch
+        {
+            ScanDataType.Short => MeetsTheScanConstraint<short>(lhs, rhs, scanConstraint),
+            ScanDataType.Integer => MeetsTheScanConstraint<int>(lhs, rhs, scanConstraint),
+            ScanDataType.Float => MeetsTheScanConstraint<float>(lhs, rhs, scanConstraint),
+            ScanDataType.Double => MeetsTheScanConstraint<double>(lhs, rhs, scanConstraint),
+            ScanDataType.Long => MeetsTheScanConstraint<long>(lhs, rhs, scanConstraint),
+            _ => throw new NotImplementedException($"Parsing string to Type: {scanConstraint.ScanDataType} not implemented")
+        };
+    }
 
-        if (!((Type)rhs.GetType()).IsValueType)
-            throw new ArgumentException("rhs must be a ValueType (struct)");
+    public static bool MeetsTheScanConstraint<T>(string lhs, string rhs, ScanConstraint scanConstraint)
+        where T : INumber<T>
+    {
+        T lhsValue = lhs.ParseNumber<T>();
+        T rhsValue = rhs.ParseNumber<T>();
 
-        return scanConstraintType switch
+        return MeetsTheScanConstraint(lhsValue, rhsValue, scanConstraint);
+    }
+
+    public static bool MeetsTheScanConstraint<T>(T lhs, T rhs, ScanConstraint scanConstraint) 
+        where T : INumber<T>
+    {
+        return scanConstraint.ScanCompareType switch
         {
             ScanCompareType.ExactValue => lhs == rhs,
             ScanCompareType.SmallerThan => lhs < rhs,
@@ -33,24 +51,29 @@ public class ValueComparer : IScanComparer
         };
     }
 
-    public IEnumerable<ProcessMemory> GetMatchingValueAddresses(IList<VirtualMemoryPage> virtualMemoryPages, IProgress<float> progressBarUpdater)
+    public IList<IMemorySegment> GetMatchingMemorySegments(IList<VirtualMemoryRegion> virtualMemoryRegions, IProgress<float>? progressBarUpdater = null)
     {
-        foreach (var virtualMemoryPage in virtualMemoryPages)
+        var matchingProcessMemories = new List<IMemorySegment>();
+
+        foreach (var virtualMemoryRegion in virtualMemoryRegions)
         {
-            for (var i = 0; i < (int)virtualMemoryPage.RegionSize; i += _sizeOfT)
+            var regionBytesAsSpan = virtualMemoryRegion.Bytes.AsSpan();
+
+            for (var i = 0; i < (int)virtualMemoryRegion.RegionSize; i += _sizeOfT)
             {
-                if (i + _sizeOfT > (int)virtualMemoryPage.RegionSize)
+                if (i + _sizeOfT > (int)virtualMemoryRegion.RegionSize)
                 {
                     break;
                 }
-                var bufferValue = virtualMemoryPage.Bytes.AsSpan().Slice(i, _sizeOfT).ToArray();
-                var valueObject = bufferValue.ByteArrayToObject(_scanConstraint.ScanDataType);
+                var memoryValue = regionBytesAsSpan.Slice(i, _sizeOfT).ConvertToString(_scanConstraint.ScanDataType);
 
-                if (CompareDataByScanConstraintType(valueObject, _userInput, _scanConstraint.ScanCompareType))
+                if (MeetsTheScanConstraint(memoryValue, _userInput, _scanConstraint))
                 {
-                    yield return new ProcessMemory(virtualMemoryPage.BaseAddress, i, bufferValue.ByteArrayToObject(_scanConstraint.ScanDataType), _scanConstraint.ScanDataType);
+                    matchingProcessMemories.Add(new MemorySegment(virtualMemoryRegion.BaseAddress, i, memoryValue, _scanConstraint.ScanDataType));
                 }
             }
         }
+
+        return matchingProcessMemories;
     }
 }
