@@ -1,10 +1,10 @@
 ﻿namespace CelSerEngine.Core.Models;
 
-public interface IProcessMemory
+public interface IMemorySegment
 {
     public IntPtr BaseAddress { get; set; }
     public int BaseOffset { get; set; }
     public IntPtr Address { get; }
-    public dynamic Value { get; set; }
+    public string Value { get; set; }
     public ScanDataType ScanDataType { get; set; }
 }
