@@ -1,22 +1,18 @@
 ﻿namespace CelSerEngine.Core.Models;
 
-public class ProcessMemory : IProcessMemory
+public class MemorySegment : IMemorySegment
 {
     public IntPtr BaseAddress { get; set; }
     public int BaseOffset { get; set; }
     public IntPtr Address => BaseAddress + BaseOffset;
-    public dynamic Value { get; set; } = 0L;
+    public string Value { get; set; }
     public ScanDataType ScanDataType { get; set; } = ScanDataType.Integer;
 
-    public ProcessMemory(IntPtr baseAddress, int baseOffset, dynamic value, ScanDataType scanDataType)
+    public MemorySegment(IntPtr baseAddress, int baseOffset, string value, ScanDataType scanDataType)
     {
         BaseAddress = baseAddress;
         BaseOffset = baseOffset;
         Value = value;
         ScanDataType = scanDataType;
-    }
-
-    public ProcessMemory(IntPtr baseAddress, int baseOffset, ScanDataType scanDataType) : this(baseAddress, baseOffset, 0, scanDataType)
-    {
     }
 }
