@@ -9,11 +9,17 @@ using System.Buffers;
 
 namespace CelSerEngine.Core.Native;
 
-public sealed class NativeApi
+// TODO: Close handles when failed, return LastErrors
+public sealed class NativeApi : INativeApi
 {
-    public static readonly ArrayPool<byte> _byteArrayPool = ArrayPool<byte>.Shared;
+    public readonly ArrayPool<byte> _byteArrayPool;
 
-    public static IntPtr OpenProcess(string processName)
+    public NativeApi()
+    {
+        _byteArrayPool = ArrayPool<byte>.Shared;
+    }
+
+    public IntPtr OpenProcess(string processName)
     {
         var processList = Process.GetProcessesByName(processName);
 
@@ -25,7 +31,7 @@ public sealed class NativeApi
         return OpenProcess(process.Id);
     }
 
-    public static IntPtr OpenProcess(int processId)
+    public IntPtr OpenProcess(int processId)
     {
         Debug.WriteLine($"ProcessID: {processId}");
 
@@ -42,8 +48,34 @@ public sealed class NativeApi
 
         return hProcess;
     }
+
+    public ProcessModuleInfo GetProcessMainModule(int processId)
+    {
+        var snapshotHandle = CreateToolhelp32Snapshot(CreateToolhelp32SnapshotFlags.TH32CS_SNAPMODULE, processId);
+
+        if (snapshotHandle == IntPtr.Zero)
+        {
+            CloseHandle(snapshotHandle);
+            throw new Exception("Failed to create snapshot.");
+        }
+
+        var moduleEntry = new MODULEENTRY32
+        {
+            dwSize = Marshal.SizeOf(typeof(MODULEENTRY32))
+        };
+
+        if (!Module32First(snapshotHandle, ref moduleEntry))
+        {
+            CloseHandle(snapshotHandle);
+            throw new Exception("Failed to get module information.");
+        }
+
+        CloseHandle(snapshotHandle);
+
+        return new ProcessModuleInfo(moduleEntry.szModule, moduleEntry.modBaseAddr, moduleEntry.modBaseSize);
+    }
     
-    public static byte[] ReadVirtualMemory(IntPtr hProcess, IntPtr address, uint numberOfBytesToRead)
+    public byte[] ReadVirtualMemory(IntPtr hProcess, IntPtr address, uint numberOfBytesToRead)
     {
         var buffer = new byte[numberOfBytesToRead];
         ReadVirtualMemory(hProcess, address, numberOfBytesToRead, buffer);
@@ -51,7 +83,7 @@ public sealed class NativeApi
         return buffer;
     }
     
-    public static void ReadVirtualMemory(IntPtr hProcess, IntPtr address, uint numberOfBytesToRead, byte[] buffer)
+    public void ReadVirtualMemory(IntPtr hProcess, IntPtr address, uint numberOfBytesToRead, byte[] buffer)
     {
         var result = NtReadVirtualMemory(
             hProcess,
@@ -64,10 +96,41 @@ public sealed class NativeApi
             throw new Exception("Failed reading memory");
     }
 
-    public static void WriteMemory(IntPtr hProcess, IProcessMemory trackedScanItem, dynamic newValue)
+    public void WriteMemory(IntPtr hProcess, IMemorySegment trackedScanItem, string newValue)
+    {
+        switch (trackedScanItem.ScanDataType)
+        {
+            case ScanDataType.Short:
+                WriteMemory(hProcess, trackedScanItem, newValue.ParseNumber<short>());
+                break;
+            case ScanDataType.Integer:
+                WriteMemory(hProcess, trackedScanItem, newValue.ParseNumber<int>());
+                break;
+            case ScanDataType.Float:
+                WriteMemory(hProcess, trackedScanItem, newValue.ParseNumber<float>());
+                break;
+            case ScanDataType.Double:
+                WriteMemory(hProcess, trackedScanItem, newValue.ParseNumber<double>());
+                break;
+            case ScanDataType.Long:
+                WriteMemory(hProcess, trackedScanItem, newValue.ParseNumber<long>());
+                break;
+            default:
+                throw new NotImplementedException($"{nameof(WriteMemory)} for Type: {trackedScanItem.ScanDataType} is not implemented");
+
+        }
+        
+    }
+
+    private void WriteMemory<T>(IntPtr hProcess, IMemorySegment trackedScanItem, T newValue)
+        where T : struct
     {
         var typeSize = trackedScanItem.ScanDataType.GetPrimitiveSize();
-        var bytesToWrite = BitConverter.GetBytes(newValue);
+        var bytesToWrite = _byteArrayPool.Rent(typeSize);
+        var ptr = Marshal.AllocHGlobal(typeSize);
+        Marshal.StructureToPtr(newValue, ptr, true);
+        Marshal.Copy(ptr, bytesToWrite, 0, typeSize);
+        Marshal.FreeHGlobal(ptr);
         var memoryAddress = trackedScanItem.Address;
 
         if (trackedScanItem is IPointer pointerItem)
@@ -81,34 +144,36 @@ public sealed class NativeApi
             bytesToWrite,
             (uint)typeSize,
             out uint _);
+
+        _byteArrayPool.Return(bytesToWrite);
     }
 
-    public static void UpdateAddresses(IntPtr hProcess, IEnumerable<IProcessMemory> virtualAddresses)
+    public void UpdateAddresses(IntPtr hProcess, IEnumerable<IMemorySegment> virtualAddresses)
     {
         foreach (var address in virtualAddresses)
         {
             if (address == null)
                 continue;
-            //ObservablePointer
-            if (address is Pointer pointerScanItem)
+
+            if (address is IPointer pointerScanItem)
             {
                 UpdatePointerAddress(hProcess, pointerScanItem);
                 continue;
             }
 
-            UpdateProcessMemory(hProcess, address);
+            UpdateMemorySegmennt(hProcess, address);
         }
     }
 
-    public static void UpdateProcessMemory(IntPtr hProcess, IProcessMemory processMemory)
+    public void UpdateMemorySegmennt(IntPtr hProcess, IMemorySegment memorySegment)
     {
-        if (processMemory == null)
+        if (memorySegment == null)
             return;
 
-        var typeSize = processMemory.ScanDataType.GetPrimitiveSize();
+        var typeSize = memorySegment.ScanDataType.GetPrimitiveSize();
         var buffer = _byteArrayPool.Rent(typeSize);
-        ReadVirtualMemory(hProcess, processMemory.Address, (uint)typeSize, buffer);
-        processMemory.Value = buffer.ByteArrayToObject(processMemory.ScanDataType);
+        ReadVirtualMemory(hProcess, memorySegment.Address, (uint)typeSize, buffer);
+        memorySegment.Value = buffer.ConvertToString(memorySegment.ScanDataType);
         _byteArrayPool.Return(buffer, clearArray: true);
     }
 
@@ -150,7 +215,7 @@ public sealed class NativeApi
     //    trackedPointerScanItem.DetermineAddressDisplayString();
     //}
 
-    public static void UpdatePointerAddress(IntPtr hProcess, Pointer? pointerAddress)
+    public void UpdatePointerAddress(IntPtr hProcess, IPointer? pointerAddress)
     {
         if (pointerAddress == null)
             return;
@@ -160,11 +225,11 @@ public sealed class NativeApi
         var typeSize = pointerAddress.ScanDataType.GetPrimitiveSize();
         var buffer = _byteArrayPool.Rent(typeSize);
         ReadVirtualMemory(hProcess, pointerAddress.PointingTo, (uint)typeSize, buffer);
-        pointerAddress.Value = buffer.ByteArrayToObject(pointerAddress.ScanDataType);
+        pointerAddress.Value = buffer.ConvertToString(pointerAddress.ScanDataType);
         _byteArrayPool.Return(buffer, clearArray: true);
     }
 
-    public static void ResolvePointerPath(IntPtr hProcess, Pointer pointerAddress)
+    public void ResolvePointerPath(IntPtr hProcess, IPointer pointerAddress)
     {
         var buffer = _byteArrayPool.Rent(sizeof(long));
         ReadVirtualMemory(hProcess, pointerAddress.Address, sizeof(long), buffer);
@@ -187,12 +252,12 @@ public sealed class NativeApi
         _byteArrayPool.Return(buffer, clearArray: true);
     }
 
-    public static IList<VirtualMemoryPage> GatherVirtualPages(IntPtr hProcess)
+    public IList<VirtualMemoryRegion> GatherVirtualMemoryRegions(IntPtr hProcess)
     {
         if (hProcess == IntPtr.Zero)
             throw new ArgumentNullException(nameof(hProcess));
 
-        var virtualMemoryPages = new List<VirtualMemoryPage>();
+        var virtualMemoryRegions = new List<VirtualMemoryRegion>();
         GetSystemInfo(out var systemInfo);
 
         // IntPtr proc_min_address = (IntPtr)0x00007ff78cb10000;
@@ -225,8 +290,8 @@ public sealed class NativeApi
                 //VirtualProtectEx(pHandle, new IntPtr((long)mem_basic_info.BaseAddress), new UIntPtr(mem_basic_info.RegionSize), 0x40, out var prt);
                 memoryBasicInfos.Add(mem_basic_info);
                 var memoryBytes = ReadVirtualMemory(hProcess, (IntPtr)mem_basic_info.BaseAddress, (uint)mem_basic_info.RegionSize);
-                var virtualMemoryPage = new VirtualMemoryPage((IntPtr)mem_basic_info.BaseAddress, mem_basic_info.RegionSize, memoryBytes);
-                virtualMemoryPages.Add(virtualMemoryPage);
+                var virtualMemoryRegion = new VirtualMemoryRegion((IntPtr)mem_basic_info.BaseAddress, mem_basic_info.RegionSize, memoryBytes);
+                virtualMemoryRegions.Add(virtualMemoryRegion);
             }
 
             // move to the next memory chunk
@@ -234,6 +299,6 @@ public sealed class NativeApi
             proc_min_address = new IntPtr((long)proc_min_address_l);
         }
 
-        return virtualMemoryPages;
+        return virtualMemoryRegions;
     }
 }
