@@ -1,31 +1,26 @@
 ﻿using System.Windows.Media;
 
-namespace CelSerEngine.Extensions
-{
-    public static class WpfExtension
-    {
-        // https://stackoverflow.com/questions/11187382/get-listview-visible-items
-        public static T? GetVisualChild<T>(this Visual referenceVisual) where T : Visual
-        {
-            Visual? child = null;
-            for (var i = 0; i < VisualTreeHelper.GetChildrenCount(referenceVisual); i++)
-            {
-                child = VisualTreeHelper.GetChild(referenceVisual, i) as Visual;
-                if (child != null && child is T)
-                {
-                    break;
-                }
-                else if (child != null)
-                {
-                    child = GetVisualChild<T>(child);
-                    if (child != null && child is T)
-                    {
-                        break;
-                    }
-                }
-            }
+namespace CelSerEngine.Wpf.Extensions;
 
-            return child as T;
+public static class WpfExtension
+{
+    // https://stackoverflow.com/questions/11187382/get-listview-visible-items
+    public static T? GetVisualChild<T>(this Visual referenceVisual) where T : Visual
+    {
+        Visual? child = null;
+        for (var i = 0; i < VisualTreeHelper.GetChildrenCount(referenceVisual); i++)
+        {
+            child = VisualTreeHelper.GetChild(referenceVisual, i) as Visual;
+            if (child != null && child is T)
+                break;
+            else if (child != null)
+            {
+                child = GetVisualChild<T>(child);
+                if (child != null && child is T)
+                    break;
+            }
         }
+
+        return child as T;
     }
 }
