@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System;
 using CelSerEngine.Core.Models;
-using System.Diagnostics.CodeAnalysis;
 
-namespace CelSerEngine.Models.ObservableModels;
+namespace CelSerEngine.Wpf.Models;
 
-public partial class ObservablePointer : ObservableProcessMemory
+public partial class ObservablePointer : ObservableMemorySegment, IPointer
 {
     [ObservableProperty]
     private string _moduleName;
@@ -14,15 +13,15 @@ public partial class ObservablePointer : ObservableProcessMemory
     [NotifyPropertyChangedFor(nameof(AddressDisplayString))]
     public IntPtr _pointingTo;
     public string ModuleNameWithBaseOffset => $"{ModuleName} + {BaseOffset:X}";
-    public List<IntPtr> Offsets { get; set; } = new List<IntPtr>();
+    public IList<IntPtr> Offsets { get; set; } = new List<IntPtr>();
     public override string AddressDisplayString => $"P->{PointingTo:X}";
 
-    public ObservablePointer(IntPtr baseAddress, int baseOffset, dynamic value, ScanDataType scanDataType) : base(baseAddress, baseOffset, 0, scanDataType)
+    public ObservablePointer(IntPtr baseAddress, int baseOffset, string value, ScanDataType scanDataType) : base(baseAddress, baseOffset, value, scanDataType)
     {
         _moduleName = "No ModuleName";
     }
 
-    public ObservablePointer(Pointer pointer) : base(pointer.BaseAddress, pointer.BaseOffset, 0, pointer.ScanDataType)
+    public ObservablePointer(Pointer pointer) : base(pointer.BaseAddress, pointer.BaseOffset, pointer.Value, pointer.ScanDataType)
     {
         _moduleName = pointer.ModuleName;
         _pointingTo = pointer.PointingTo;
