@@ -1,7 +1,6 @@
-﻿using CelSerEngine.Models.ObservableModels;
-using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
 
-namespace CelSerEngine.Models;
+namespace CelSerEngine.Wpf.Models;
 
 public partial class TrackedItem : ObservableObject
 {
@@ -10,15 +9,15 @@ public partial class TrackedItem : ObservableObject
     [ObservableProperty]
     private string _description;
 
-    public ObservableProcessMemory Item { get; set; }
-    public dynamic? SetValue { get; set; }
+    public ObservableMemorySegment Item { get; set; }
+    public string? SetValue { get; set; }
 
     partial void OnIsFreezedChanged(bool value)
     {
         SetValue = value ? Item.Value : null;
     }
 
-    public TrackedItem(ObservableProcessMemory item)
+    public TrackedItem(ObservableMemorySegment item)
     {
         Item = item;
         _description = "Description";
