@@ -1,7 +1,6 @@
-﻿using CelSerEngine.ViewModels;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 
-namespace CelSerEngine;
+namespace CelSerEngine.Wpf.ViewModels;
 
 public class ViewModelLocator
 {
